% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\docType{data}
\name{CL}
\alias{CL}
\title{A cluster configuration of spheroids}
\format{A list of length 9. Each element contains a list of spheroids}
\usage{
data(AL2MC_20p_k10_F2p_CL)
}
\description{
A dataset containing cluster regions of spheroids
where each list element consists of a spheroids list
with additional attributes \code{class="prolate"},
\code{interior} (logical), indicating whether the cluster
fully lies in the interior of the simulation box
}
\author{
M. Baaske
}
\keyword{datasets}
