% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\docType{data}
\name{CLF}
\alias{CLF}
\title{A cluster configuration of fibers}
\format{A list of length 11 containing lists of fibers belonging to the
		   corresponding cluster}
\usage{
data(AL2MC_15p_k10_F2p_CL)
}
\description{
A dataset containing clusters of regions of fibers,
including a second phase as spheres where each list element consists
of a list of fibers with additional attributes \code{class="cylinder"},
\code{interior} (logical), indicating whether the cluster fully lies in
the interior of the simulation box.
}
\author{
M. Baaske
}
\keyword{datasets}
