% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\name{drawProjection}
\alias{drawProjection}
\title{Draw defect projections of particles}
\usage{
drawProjection(E, conv = TRUE, np = 25)
}
\arguments{
\item{E}{a list of ellipses (the projected particles)}

\item{conv}{logical, if \code{conv=TRUE} (default) the convex hull is drawn}

\item{np}{number of points used for sampling the convex hull of projections}
}
\value{
draw projections and the convex hull (if enabled) in a 3d plot
		   return area of polygon and points of convex hull
}
\description{
The function draws the defect projections in a 3d plot after \code{spheroids3d}
has been called.
}
\author{
M. Baaske
}
