% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\name{extractClusters}
\alias{extractClusters}
\title{Extract defects}
\usage{
extractClusters(clust, area)
}
\arguments{
\item{clust}{list of defects, see \code{\link{simDefect}}}

\item{area}{the area value lower bound}
}
\value{
list of clusters
}
\description{
Extract all defects with some area value larger than \code{area}
}
\details{
The function extracts all defects (convex hulls of projected defect areas)
which have a larger area value than the predefined \code{area} value for some
further analysis.
}
\author{
M. Baaske
}
