% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\name{getSphereProjection}
\alias{getSphereProjection}
\title{Sphere projections}
\usage{
getSphereProjection(S, draw = TRUE, conv = TRUE, np = 20)
}
\arguments{
\item{S}{particles to be projected}

\item{draw}{logical, \code{draw=TRUE} (default) draw the projected spheres}

\item{conv}{logical, \code{conv=TRUE} (default) draw the convex hull}

\item{np}{number of points used for sampling the convex hull of projections}
}
\value{
draw projections and the convex hull (if enabled) in a 3d plot
		   returning its polygonal area and points of the convex hull
}
\description{
The function draws the defect projections in a 3d plot
and returns the points of its  the convex hull together
with points of each projected particle (here a sphere).
}
\author{
M. Baaske
}
