% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\name{getSpheroidProjection}
\alias{getSpheroidProjection}
\title{Projection of defect types}
\usage{
getSpheroidProjection(S, type)
}
\arguments{
\item{S}{list of (non-overlapping) particles (spheroids)}

\item{type}{either \code{crack} or \code{delam}}
}
\value{
returns a list of 2d ellipse objects.
}
\description{
In case one wishes to analyze the defect projections directly the function returns
the projected objects itself. The two types of defects lead to different projected objects.
In case of defect type \code{crack} the circle of maximum radius of
the particle is orthogonally projected otherwise the whole particle
is projected in the main load direction, i.e. the xy plane.
}
\details{
Get the projected objects (ellipses) of particles dependant on their defect types
}
\author{
M. Baaske
}
