% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\docType{data}
\name{S}
\alias{S}
\title{A packed spheroids configuration (particle system)}
\format{A list of length 4905 containing lists each of length six}
\usage{
data(AL2MC_20p_k10_F2p_S)
}
\description{
A dataset containing non-overlapping spheroids
where each list element consists of the following variables:
}
\details{
\itemize{
	 \item{id}{ the current cluster id}
	 \item{center}{ the center of ball which defines the cluser region}
	 \item{u}{ the orientation vector of spheroid's rotational symmetry axis}
	 \item{acb}{ two shorter semi-axis lengths and major semi-axis length}
	 \item{angles}{ polar angle and azimuthal angle}
	 \item{rotM}{ rotational matrix, center of rotation is the center of the spheroid}
	 \item{interior}{ attribute, whether the spheroid lies in the interior of the simulation box}
	 \item{label}{ attribute, character, user defined label, here: 'P' (particle) or 'F' (ferrit)}
  \item{area}{ attribute, the projected area}
}
}
\author{
M. Baaske
}
\keyword{datasets}
