% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\docType{data}
\name{SF}
\alias{SF}
\title{A packed (sphero)cylinder configuration (fiber system)}
\format{A list of length 3550 containing lists each of length nine}
\usage{
data(AL2MC_15p_k10_F2p_S)
}
\description{
A dataset containing non-overlapping spherocylinders
where each list element consists of the following variables:
}
\details{
\itemize{
	 \item{id}{ the current cluster id}
	 \item{center}{ the center of ball which defines the cluser region}
	 \item{origin0}{ the center of the fist cylinder cap}
	 \item{origin1}{ the center of the second cylinder cap}
  \item{length}{ length/height of cylinder without radii of caps}
	 \item{u}{ orientation vector of spheroid's rotational symmetry axis}
	 \item{r}{ radius of cylinder}
	 \item{angles}{ colatitude and longitude angle}
	 \item{rotM}{ rotational matrix, center of rotation is the center of the spheroid}
	 \item{radius}{ attribute, random radius of perfect simulation, if enabled}
	 \item{interior}{ attribute, whether the spheroid lies in the interior of the simulation box}
	 \item{label}{ attribute, character, user defined label, here: 'P' (particle) or 'F' (ferrit)}
  \item{area}{ attribute, the projected area}
}
}
\author{
M. Baaske
}
\keyword{datasets}
