% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\name{drawDefectProjections}
\alias{drawDefectProjections}
\title{Draw defect projections}
\usage{
drawDefectProjections(F, D, ...)
}
\arguments{
\item{F}{geometry system}

\item{D}{defects object}

\item{...}{further arguments passed to
\code{\link{getSphereProjection}},
 \code{\link{getSpheroidProjection}},
 \code{\link{getCylinderProjection}}}
}
\value{
\code{NULL}
}
\description{
Draw defect projections for spheroids, spherocylinders or spheres.
The method first constructs the projected objects sampling some points
on the borders and calculates the convex hull of such points.
The defect object as returned by \code{\link{simFracture}}
and used to show the accumulated defects.
}
\author{
M. Baaske
}
