% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{simCrackTime}
\alias{simCrackTime}
\title{Defect failure times}
\usage{
simCrackTime(S, stress, vickers, param, fun = lapply)
}
\arguments{
\item{S}{geometry objects system}

\item{stress}{the applied stress level}

\item{vickers}{Vickers hardness, see details}

\item{fun}{optional, either \code{lapply} (default) or parllel processing by \code{mclapply}}
}
\value{
a list with the following elements:
			\itemize{
				\item{id}{ id of particle }
				\item{U}{ crack failure time }
				\item{V}{ delamination failure time }
				\item{T}{ the minimum of both failure times}
				\item{B}{ failure type, either 0  for particle crack or 1 for particle delamination }
				\item{A}{ projection area set to zero for later use }
				\item{label}{ either \code{label="P"} for primary phase or \code{label="F"}
							  for secondary phase }
			}
}
\description{
Simulation of individual defect failure times. For a secondary phase only
the defect type "delamination" is considered.
}
\author{
Felix Ballani, Markus Baaske
}
