% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLife.R
\name{woehlerDiagram}
\alias{woehlerDiagram}
\title{Plot Woehler diagram}
\usage{
woehlerDiagram(W, W2 = NULL, yrange = NULL, main = "",
  legend.text = NULL, ...)
}
\arguments{
\item{W}{results from function \code{\link{woehler}}}

\item{W2}{optional, \code{W2=NULL} (default) woheler results}

\item{yrange}{range of stress levels, the min/max values of stress levels (default)}

\item{main}{main title of the plot}

\item{legend.text}{text for the legend as vector of strings, \code{NULL} (default) for no legend}

\item{...}{graphic parameters passed \code{legend}}
}
\value{
\code{NULL}
}
\description{
Plot a Woehler diagram for simulated Woehler experiments
}
\details{
The simulated results of the Woehler experiments are summarized in
the load-cycle diagram as usually done in the Woehler diagram.
}
\author{
M. Baaske
}
